requirements.txt
beautifulsoup4
pyyaml
sqlitedict